fix/refactor and codeclimate
