+++ src/services/requestAPI.js
@@ -1,10 +1,10 @@
-const requestAPI = (url) =>
-  fetch(url).then((response) =>
-    response
-      .json()
-      .then((json) =>
-        response.ok ? Promise.resolve(json) : Promise.reject(json),
-      ),
-  );
+// const requestAPI = (url) =>
+//   fetch(url).then((response) =>
+//     response
+//       .json()
+//       .then((json) =>
+//         response.ok ? Promise.resolve(json) : Promise.reject(json),
+//       ),
+//   );
 
-export default requestAPI;
+// export default requestAPI;
